refactor(rust-api): extract job + worker handlers into modules (sc-1636)
Move the job lifecycle handlers (list/create/claim/get/cancel/retry/
duplicate job, update_job_progress + the asset-persist and trained-LoRA
registration helpers) into jobs.rs, and the worker/queue handlers
(queue_summary, list_workers, person readiness, register/heartbeat
worker) into workers.rs. The event SSE handlers and event structs stay
at the root for a future events module.

Same pattern: `use super::*;` + pub(crate). tests.rs's
person_readiness_from_workers import is repointed to super::workers.
Behavior-preserving. lib.rs 6,407 -> 5,938 lines.
